refactor: modularize crag layer
